microservicios/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
PyJWT==2.10.1
requests==2.32.4
Werkzeug==3.0.3
bcrypt==4.1.2
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.0.0